# docs/swagger.yaml
basePath: /
definitions:
  doc.Chat:
    description: All post information
    properties:
      chat:
        $ref: '#/definitions/entities.Chat'
    type: object
  doc.ChatCreateRequest:
    description: All post information
    properties:
      body:
        $ref: '#/definitions/dto.CreateChatDTO'
    type: object
  doc.ChatCreateResponse:
    description: All post information
    properties:
      body:
        $ref: '#/definitions/doc.Chat'
    type: object
...
  doc.Exists:
    properties:
      has_dialog:
        type: boolean
    type: object
  doc.HasDialogResponse:
    description: All post information
    properties:
      body:
        $ref: '#/definitions/doc.Exists'
    type: object
  doc.Messages:
    description: All post information
    properties:
      messages:
        items:
          $ref: '#/definitions/entities.Message'
        type: array
    type: object
  doc.MessagesListResponse:
    description: All post information
    properties:
      body:
        $ref: '#/definitions/doc.Messages'
    type: object
...
  doc.SendMessageResponse:
    description: All post information
    properties:
      body:
        $ref: '#/definitions/dto.NewMessage'
    type: object
...
  dto.CreateChatDTO:
    properties:
      user_links:
        items:
          type: string
        type: array
    type: object
...
  dto.GetChatsDTO:
    properties:
      batch_size:
        type: integer
      offset:
        type: integer
    type: object
  dto.NewMessage:
    properties:
      chat_id:
        type: integer
      message_content_type:
        type: string
      reply_to:
        type: integer
      text_content:
        type: string
    type: object
...
  entities.Chat:
    properties:
      chat_id:
        type: integer
      user_links:
        items:
          type: string
        type: array
    type: object
...
  entities.Message:
    properties:
      change_date:
        type: string
      chat_id:
        type: integer
      creation_date:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      link:
        type: string
      message_content_type:
        type: string
      reply_to:
        type: integer
      text_content:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: CashMap team
  description: Api for Depeche social network. VK Education project. Spring, 2023
  title: Depeche API
  version: "1.0"
paths:
...
  /api/im/chat/check:
    get:
      description: User can check if chat with that user_link exists
      parameters:
      - description: User link
        in: query
        name: user_link
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.HasDialogResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Check if dialog exists
      tags:
      - Message
  /api/im/chat/create:
    post:
      description: Create new chat with user
      parameters:
      - description: Chat info
        in: body
        name: request
        schema:
          $ref: '#/definitions/doc.ChatCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.ChatCreateResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Create chat
      tags:
      - Message
  /api/im/chats:
    get:
      description: Get chats list
      parameters:
      - description: Batch size
        in: query
        name: batch_size
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.GetChatsDTO'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Get chats
      tags:
      - Message
  /api/im/messages:
    get:
      description: Get messages batch by chatID sorted by date
      parameters:
      - description: Chat id
        in: query
        name: chat_id
        type: integer
      - description: Batch size
        in: query
        name: batch_size
        type: integer
      - description: Last post date
        in: query
        name: last_post_date
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.MessagesListResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Get messages
      tags:
      - Message
  /api/im/send:
    post:
      description: Add message to db and send it to listeners
      parameters:
      - description: Message info
        in: body
        name: request
        schema:
          $ref: '#/definitions/doc.SendMessageResponse'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Send message
      tags:
      - Message
...
swagger: "2.0"